some simple docs

+++ tensorpack/dataflow/dataset/ilsvrc.py
@@ -79,7 +79,7 @@ class ILSVRCMeta(object):
         mean_file = os.path.join(self.dir, 'imagenet_mean.binaryproto')
         with open(mean_file) as f:
             obj.ParseFromString(f.read())
-        arr = np.array(obj.data).reshape((3, 256, 256))
+        arr = np.array(obj.data).reshape((3, 256, 256)).astype('float32')
         arr = np.transpose(arr, [1,2,0])
         if size is not None:
             arr = cv2.resize(arr, size[::-1])
